# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "STX API",
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/v2-stx-api/public"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Portal"
                ],
                "operationId": "0c8497c027312a339b682a169654c4bb",
                "parameters": [
                    {
                        "name": "username",
                        "in": "query",
                        "description": "Username of login portal user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password of login portal user",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "password"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login portal"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Portal"
                ],
                "operationId": "c48c74b1cf1afca4347d7ceee2c29088",
                "parameters": [
                    {
                        "name": "username",
                        "in": "query",
                        "description": "Username of login portal user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email of login portal user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password of login portal user",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "password"
                        }
                    },
                    {
                        "name": "confirm_password",
                        "in": "query",
                        "description": "Password Confirmation of login portal user",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "password"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login portal"
                    }
                }
            }
        },
        "/api/portal/users": {
            "get": {
                "tags": [
                    "Portal"
                ],
                "operationId": "e21a6149608ec1065103b3be108f642f",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Auth token from login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Username",
                        "in": "header",
                        "description": "Username from login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get list of all users"
                    }
                }
            }
        }
    }
}